requirements.txt
fastapi
uvicorn
aiohttp
pydantic
python-dotenv